requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.31.1
httpx==0.27.2
pydantic>=2.11.1,<3.0.0
python-dotenv==1.0.1
# Optional MCP integration for later use in the orchestrator
mcp>=1.21.0
